Update changes made to handle file paths

+++ hds_functions/json_utils.py
@@ -5,9 +5,11 @@ Description:
     This module provides utilities for reading and writing JSON files in Python.
 """
 import json
+import os
 from typing import Any, Dict
+from .environment_utils import resolve_path
 
-def read_json_file(path: str) -> Dict[str, Any]:
+def read_json_file(path: str, repo: str = None) -> Dict[str, Any]:
     """
     Read a JSON file from the specified path and return its content as a Python dictionary. 
 
@@ -15,7 +17,8 @@ def read_json_file(path: str) -> Dict[str, Any]:
     It checks for duplicate keys within the JSON object and raises a ValueError if duplicates are found.
 
     Args:
-        path (str): The path to the JSON file.
+        path (str): Path to the JSON file. Can be an absolute path, a relative path with './', or a path within a repo.
+        repo (str): Name of the repo if the path is relative within a repo.
 
     Returns:
         dict: A Python dictionary containing the JSON content.
@@ -24,9 +27,9 @@ def read_json_file(path: str) -> Dict[str, Any]:
         ValueError: If the JSON file contains duplicate keys.
 
     Examples:
-        >>> data = read_json_file("example.json")
-        >>> print(data)
-        {'key1': 'value1', 'key2': 'value2'}
+        >>> data = read_json_file('./relative/path/in/project.json')
+        >>> data = read_json_file('/Workspace/absolute/path.json')
+        >>> data = read_json_file(path='path/in/repo.json', repo='common_repo')
     """
 
     def check_json_for_duplicate_keys(ordered_pairs):
@@ -45,19 +48,22 @@ def read_json_file(path: str) -> Dict[str, Any]:
         d = {}
         for k, v in ordered_pairs:
             if k in d:
-                raise ValueError(f"JSON file '{path}' contains duplicate key: {k}")
+                raise ValueError(f"JSON file '{resolved_path}' contains duplicate key: {k}")
             else:
                 d[k] = v
         return d
+
+    # Resolve the path
+    resolved_path = resolve_path(path, repo)
     
     # Load JSON file and check for duplicate keys
-    with open(path) as json_file:
+    with open(resolved_path) as json_file:
         json_dict = json.load(json_file, object_pairs_hook=check_json_for_duplicate_keys)
 
     return json_dict
 
 
-def write_json_file(data: Dict[str, Any], path: str, indent: int = 4) -> None:
+def write_json_file(data: Dict[str, Any], path: str, repo: str = None, indent: int = 4) -> None:
     """
     Write a Python dictionary to a JSON file at the specified path.
 
@@ -66,7 +72,8 @@ def write_json_file(data: Dict[str, Any], path: str, indent: int = 4) -> None:
 
     Args:
         data (dict): The Python dictionary to be written to the JSON file.
-        path (str): The path to the JSON file.
+        path (str): Path to the JSON file. Can be an absolute path, a relative path with './', or a path within a repo.
+        repo (str): Name of the repo if the path is relative within a repo.
         indent (int, optional): The number of spaces used for indentation (default is 4).
 
     Returns:
@@ -74,7 +81,18 @@ def write_json_file(data: Dict[str, Any], path: str, indent: int = 4) -> None:
 
     Examples:
         >>> data = {'key1': 'value1', 'key2': 'value2'}
-        >>> write_json_file(data, "output.json")
+        >>> write_json_file(data, "./in_project_folder.json")
+        >>> write_json_file(data, "/Workspace/absolute_path.json")
+        >>> write_json_file(data, path="in/shared/repo.json", repo="common_repo")
     """
+    # Resolve the path
+    resolved_path = resolve_path(path, repo)
+
+    # Check if the directory exists
+    directory = os.path.dirname(resolved_path)
+    if not os.path.exists(directory):
+        raise ValueError(f"Directory '{directory}' does not exist.")
+
+    # Write file
     with open(path, 'w') as fp:
         json.dump(data, fp, indent=indent)
